implemented tracker in orc

+++ pkg/user/user_input.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Reps struct {
-	Weight int `json:"weight"`
+	Weight float32 `json:"weight"`
 	Reps   int `json:"reps"`
 }
