+++ services/tracker_service/migrations/000002_create_workout_table.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE workout (
 	tracker_id UUID NOT NULL,
 	exercise_id UUID NOT NULL,
 	set_number INTEGER NOT NULL,
-	weight INTEGER,
+	weight DECIMAL(10, 2),
 	reps INTEGER,
 	FOREIGN KEY (tracker_id) REFERENCES tracker(id)
 )
